bugfix: if etcd is standalone, these commands stumble as the dirs do not exist

+++ providers/openshift_redeploy_certificate.rb
@@ -25,18 +25,22 @@ action :redeploy do
   execute 'Backup master stuff' do
     command 'tar czvf master-backup-$(date +%s).tar.gz /etc/origin/master /var/www/html/master --ignore-failed-read && rm -rf /var/www/html/master'
     cwd '/etc/origin'
+    only_if '[ -a /etc/origin/master ]'
   end
   execute 'Backup node stuff' do
     command 'tar czvf node-backup-$(date +%s).tar.gz /etc/origin/node /var/www/html/node --ignore-failed-read --remove-files'
     cwd '/etc/origin'
+    only_if '[ -a /etc/origin/node ]'
   end
   execute 'Delete old certs' do
     command 'rm -f $(ls -I serviceaccounts\* -I registry\*)'
     cwd '/etc/origin/master'
+    only_if '[ -a /etc/origin/master ]'
   end
   execute 'Remove root kubeconfig' do
     command 'rm -f config'
     cwd '/root/.kube'
+    only_if '[ -a /root/.kube ]'
   end
   include_recipe 'cookbook-openshift3::default'
   new_resource.updated_by_last_action(true)
